requirements.txt
Django==4.2.7
djangorestframework==3.14.0
gunicorn==20.1.0
whitenoise==6.4.0
